Update to docs. Small updates to code: * JSGame.cpp - make getTradeInfo return something in the currently disabled     nMode == 1 * JSUnit.cpp - make getMerc return null if no merc is found * JSCore.cpp - if load(scriptname) fails, return null instead of false. Done     because function now returns a Script instead of

+++ doc/Party.h
@@ -18,7 +18,7 @@ public:
 
 	/** The unique id of the RosterUnit.
 	 */
-	int gid;
+	double gid;
 
 	/** The life of the RosterUnit.
 	 *
